+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [v1.0.2]
+## [v1.0.2] - 2020-12-29
 
 ### Added
 - Set "kv-user" cookie on sign-in and registration